requirements.txt
sqlalchemy
requests
opentelemetry-instrumentation>=0.30b1
opentelemetry-instrumentation-sqlalchemy>=0.30b1
opentelemetry-sdk>=1.11.1
opentelemetry-api>=1.11.1
opentelemetry-semantic-conventions>=0.30b1
opentelemetry-util-http>=0.30b1
sentry_sdk~=1.9
